refactor: extract prText to its own file
